Move MockView to SplitTreeTests itself
