--- macos/Tests/Helpers/MockView.swift
@@ -1,27 +1,0 @@
-import AppKit
-
-// @preconcurrency suppresses Sendable errors from Codable on NSView
-// but the Swift compiler still complains about it.
-class MockView: NSView, @preconcurrency Codable, Identifiable {
-    let id: UUID
-
-    init(id: UUID = UUID()) {
-        self.id = id
-        super.init(frame: .zero)
-    }
-
-    required init?(coder: NSCoder) { fatalError() }
-
-    enum CodingKeys: CodingKey { case id }
-
-    required init(from decoder: Decoder) throws {
-        let c = try decoder.container(keyedBy: CodingKeys.self)
-        self.id = try c.decode(UUID.self, forKey: .id)
-        super.init(frame: .zero)
-    }
-
-    func encode(to encoder: Encoder) throws {
-        var c = encoder.container(keyedBy: CodingKeys.self)
-        try c.encode(id, forKey: .id)
-    }
-}
