+++ macos/Tests/Splits/SplitTreeTests.swift
@@ -2,6 +2,30 @@ import AppKit
 import Testing
 @testable import Ghostty
 
+class MockView: NSView, Codable, Identifiable {
+    let id: UUID
+
+    init(id: UUID = UUID()) {
+        self.id = id
+        super.init(frame: .zero)
+    }
+
+    required init?(coder: NSCoder) { fatalError() }
+
+    enum CodingKeys: CodingKey { case id }
+
+    required init(from decoder: Decoder) throws {
+        let c = try decoder.container(keyedBy: CodingKeys.self)
+        self.id = try c.decode(UUID.self, forKey: .id)
+        super.init(frame: .zero)
+    }
+
+    func encode(to encoder: Encoder) throws {
+        var c = encoder.container(keyedBy: CodingKeys.self)
+        try c.encode(id, forKey: .id)
+    }
+}
+
 struct SplitTreeTests {
     /// Creates a two-view horizontal split tree (view1 | view2).
     private func makeHorizontalSplit() throws -> (SplitTree<MockView>, MockView, MockView) {
